Day 5
Got stuck for a while on part1 with what turned out to be a bug with the parsing and neglecting to include the final category. Unfortunately, the sample test scenario doesn't make use of any mappings in the final block.

P2 wasn't able to make use of the logic from p1, but luckily that logic was fairly small relative to the parsing effort. P2 was mostly a challenge of off-by-1 errors and the logic of comparing and splitting the the different ranges.