Day 7
A fun one, spent the vast majority of time on p1 and was glad p2 was fairly easily handled by passing a `with_jokers` bool down the chain. Was pleased with the outcome of implementing Ord for the Hand struct which made for very readable logic in the main p1 and p2 fns.